REF: .drop tests (#33156)
